==> 4_redo_embeddings.py <==
from typing import List, Optional
import pandas as pd
from vertexai.language_models import TextEmbeddingInput, TextEmbeddingModel
import time

### Code related to Vertex AI (text embedding) from "https://cloud.google.com/vertex-ai/generative-ai/docs/embeddings/get-text-embeddings#get_text_embeddings_for_a_snippet_of_text"

df = pd.read_excel("emb_videos.xlsx")

## Function to embed a text (taken from the VERTEX documentation and adapted)
def embed_text(
    texts: List[str] = ["banana muffins? ", "banana bread? banana muffins?"],
    task: str = "RETRIEVAL_DOCUMENT",
    model_name: str = "text-embedding-004",
    dimensionality: Optional[int] = 256,
) -> List[List[float]]:
    """Embeds texts with a pre-trained, foundational model."""
    model = TextEmbeddingModel.from_pretrained(model_name)
    inputs = [TextEmbeddingInput(text, task) for text in texts]
    kwargs = dict(output_dimensionality=dimensionality) if dimensionality else {}
    embeddings = model.get_embeddings(inputs, **kwargs)
    return [embedding.values for embedding in embeddings]



counter = 0
worked = 0

for iteration in range(3):

    print(f"\n###### Iteration {iteration}\n")

    for row in df.itertuples():

        # print(row.captions)
        counter += 1

        if counter % 1000 == 0: # just for me to know over how many entries it iterated
                print("1000 done")

        if row.captions == False or row.captions == "[]": # if the value of captions is merely "[]" or False, it is skipped
            df.at[row.Index, 'embeddings'] = False
            #  print("skipped")
            continue
        
        if row.embeddings != False: # this is the case, if there is already an embedding from before
            continue # skipping to the next

        try:
            # print(embed_text([row.captions]))
            test = embed_text([row.captions])
            # print(len(test[0]))
            df.at[row.Index, 'embeddings'] = str(test[0]) # has to be converted to string, otherwise getting an error when saving
            # print("worked!")
            worked += 1
            time.sleep(0.003) # to make sure no number of requests per minute is exceeded
        except:
            # print("Embedding generation not working")
            df.at[row.Index, 'embeddings'] = False

        

        

    print(f"\n###### Finished Iteration {iteration}\n")
    

# print(df)


# # saving the df
df.to_excel('red_emb_videos.xlsx', index=False)

print(worked)
